scripts/requirements.txt
docutils==0.16
lockfile==0.12.2
psutil==5.7.0
python-daemon==2.2.4
websockets==8.1
